Fold CSS modules exports example and scoping notes into examples/css
Merge the separate css-module-exports example into the canonical css
example and move the scoping reference out of lib into that example's
README (template). One CSS example now shows native CSS, the
getJSON-equivalent name-map plugin, and what native CSS scopes.
