--- examples/css-module-exports/build.js
@@ -1,1 +1,0 @@
-require("../build-common");
