--- examples/css-module-exports/style.module.css
@@ -1,8 +1,0 @@
-.header {
-	color: rebeccapurple;
-}
-
-.link {
-	composes: header;
-	text-decoration: underline;
-}
