--- examples/css-module-exports/template.md
@@ -1,46 +1,0 @@
-# example.js
-
-```javascript
-_{{example.js}}_
-```
-
-# style.module.css
-
-```css
-_{{style.module.css}}_
-```
-
-# webpack.config.js
-
-```javascript
-_{{webpack.config.js}}_
-```
-
-# dist/output.css
-
-```css
-_{{dist/output.css}}_
-```
-
-# dist/style.module.css.json
-
-The CSS Modules name map, emitted by the plugin — the same data postcss-modules
-exposes via its `getJSON` callback.
-
-```json
-_{{dist/style.module.css.json}}_
-```
-
-# Info
-
-## Unoptimized
-
-```
-_{{stdout}}_
-```
-
-## Production mode
-
-```
-_{{production:stdout}}_
-```
